Moved item and terrain-specific actions to the appropriate namespaces
