+++ SolStandard/Entity/Unit/Actions/Item/DeployBarricadeAction.cs
@@ -7,7 +7,7 @@ using SolStandard.Map.Elements.Cursor;
 using SolStandard.Utility.Assets;
 using SolStandard.Utility.Events;
 
-namespace SolStandard.Entity.Unit.Actions
+namespace SolStandard.Entity.Unit.Actions.Item
 {
     public class DeployBarricadeAction : UnitAction
     {
