+++ SolStandard/Entity/Unit/Actions/Item/DeployBombAction.cs
@@ -8,7 +8,7 @@ using SolStandard.Map.Elements.Cursor;
 using SolStandard.Utility.Assets;
 using SolStandard.Utility.Events;
 
-namespace SolStandard.Entity.Unit.Actions
+namespace SolStandard.Entity.Unit.Actions.Item
 {
     public class DeployBombAction : UnitAction
     {
