+++ SolStandard/Entity/Unit/Actions/Item/DeployRecallPointAction.cs
@@ -10,7 +10,7 @@ using SolStandard.Map.Elements.Cursor;
 using SolStandard.Utility.Assets;
 using SolStandard.Utility.Events;
 
-namespace SolStandard.Entity.Unit.Actions
+namespace SolStandard.Entity.Unit.Actions.Item
 {
     public class DeployRecallPointAction : UnitAction
     {
