+++ SolStandard/Entity/Unit/Actions/Terrain/VendorPurchase.cs
@@ -11,7 +11,7 @@ using SolStandard.Map.Elements.Cursor;
 using SolStandard.Utility.Assets;
 using SolStandard.Utility.Events;
 
-namespace SolStandard.Entity.Unit.Actions
+namespace SolStandard.Entity.Unit.Actions.Terrain
 {
     public class VendorPurchase : UnitAction
     {
