+++ SolStandard/SolStandard.csproj
@@ -130,8 +130,6 @@
     <Compile Include="Entity\ITriggerable.cs" />
     <Compile Include="Entity\Unit\Actions\Archer\HuntingTrap.cs" />
     <Compile Include="Entity\Unit\Actions\Archer\PoisonArrow.cs" />
-    <Compile Include="Entity\Unit\Actions\BankDeposit.cs" />
-    <Compile Include="Entity\Unit\Actions\BankWithdraw.cs" />
     <Compile Include="Entity\Unit\Actions\Bard\Crescendo.cs" />
     <Compile Include="Entity\Unit\Actions\Champion\Challenge.cs" />
     <Compile Include="Entity\Unit\Actions\Champion\Shove.cs" />
@@ -149,12 +147,8 @@
     <Compile Include="Entity\Unit\Actions\Creeps\TreasureHunterRoutine.cs" />
     <Compile Include="Entity\Unit\Actions\Creeps\TriggerHappyRoutine.cs" />
     <Compile Include="Entity\Unit\Actions\Creeps\WaitRoutine.cs" />
-    <Compile Include="Entity\Unit\Actions\DeployBarricadeAction.cs" />
     <Compile Include="Entity\Unit\Actions\Creeps\WanderRoutine.cs" />
     <Compile Include="Entity\Unit\Actions\DamageOverTimeAction.cs" />
-    <Compile Include="Entity\Unit\Actions\DeployBombAction.cs" />
-    <Compile Include="Entity\Unit\Actions\DeployLadderBridgeAction.cs" />
-    <Compile Include="Entity\Unit\Actions\DeployRecallPointAction.cs" />
     <Compile Include="Entity\Unit\Actions\DraftNewUnitAction.cs" />
     <Compile Include="Entity\Unit\Actions\DropGiveGoldAction.cs" />
     <Compile Include="Entity\Unit\Actions\Duelist\Focus.cs" />
@@ -162,6 +156,12 @@
     <Compile Include="Entity\Unit\Actions\Duelist\Shift.cs" />
     <Compile Include="Entity\Unit\Actions\Guard.cs" />
     <Compile Include="Entity\Unit\Actions\ICombatProc.cs" />
+    <Compile Include="Entity\Unit\Actions\Item\DeployBarricadeAction.cs" />
+    <Compile Include="Entity\Unit\Actions\Item\DeployBombAction.cs" />
+    <Compile Include="Entity\Unit\Actions\Item\DeployLadderBridgeAction.cs" />
+    <Compile Include="Entity\Unit\Actions\Item\DeployRecallPointAction.cs" />
+    <Compile Include="Entity\Unit\Actions\Item\MagneticPullAction.cs" />
+    <Compile Include="Entity\Unit\Actions\Item\ReturnToRecallPointAction.cs" />
     <Compile Include="Entity\Unit\Actions\ITurnProc.cs" />
     <Compile Include="Entity\Unit\Actions\IIncrementableAction.cs" />
     <Compile Include="Entity\Unit\Actions\IRoutine.cs" />
@@ -177,7 +177,6 @@
     <Compile Include="Entity\Unit\Actions\Mage\Ignite.cs" />
     <Compile Include="Entity\Unit\Actions\Mage\Inferno.cs" />
     <Compile Include="Entity\Unit\Actions\Mage\Replace.cs" />
-    <Compile Include="Entity\Unit\Actions\MagneticPullAction.cs" />
     <Compile Include="Entity\Unit\Actions\Marauder\Brace.cs" />
     <Compile Include="Entity\Unit\Actions\Marauder\Guillotine.cs" />
     <Compile Include="Entity\Unit\Actions\Marauder\Rage.cs" />
@@ -186,12 +185,13 @@
     <Compile Include="Entity\Unit\Actions\Pugilist\Meditate.cs" />
     <Compile Include="Entity\Unit\Actions\Pugilist\PressurePoint.cs" />
     <Compile Include="Entity\Unit\Actions\Pugilist\Uppercut.cs" />
-    <Compile Include="Entity\Unit\Actions\ReturnToRecallPointAction.cs" />
     <Compile Include="Entity\Unit\Actions\SpawnUnitAction.cs" />
+    <Compile Include="Entity\Unit\Actions\Terrain\BankDeposit.cs" />
+    <Compile Include="Entity\Unit\Actions\Terrain\BankWithdraw.cs" />
     <Compile Include="Entity\Unit\Actions\Terrain\EscapeAction.cs" />
     <Compile Include="Entity\Unit\Actions\Terrain\TradeItemAction.cs" />
-    <Compile Include="Entity\Unit\Actions\VendorPurchase.cs" />
     <Compile Include="Entity\Unit\Actions\Sprint.cs" />
+    <Compile Include="Entity\Unit\Actions\Terrain\VendorPurchase.cs" />
     <Compile Include="Entity\Unit\CreepEntity.cs" />
     <Compile Include="Entity\Unit\CreepUnit.cs" />
     <Compile Include="Entity\Unit\GameUnit.cs" />
